Extract sensor rotation shape calculator
