+++ app/src/main/java/com/skul/yuriy/composeplayground/feature/sensorRotation/shape/RotationShapeCalculatorContract.kt
@@ -1,0 +1,13 @@
+package com.skul.yuriy.composeplayground.feature.sensorRotation.shape
+
+import androidx.compose.ui.geometry.Offset
+
+interface RotationShapeCalculatorContract {
+    fun calculate(
+        anchorA: Offset,
+        anchorB: Offset,
+        anchorC: Offset,
+        anchorD: Offset,
+        rotationDegrees: Float
+    ): ShapePoints
+}
